yosi

+++ src/diary/views.py
@@ -557,9 +557,11 @@ def week_graph(request,selected_date=None):
         # 総評コメントを取得
         week_ai=Week_AI.objects.get(user = request.user,created_date__range=[start_date,one_week_str])
         ai_comment = week_ai.ai_comment
-    else:
+    elif week_ai:
       week_ai=Week_AI.objects.get(user = request.user,created_date__range=[start_date,one_week_str])
       week_ai.delete()
+      ai_comment = '4日以上日記をかいてくにゃさい'
+    else:
       ai_comment = '4日以上日記をかいてくにゃさい'
     data = chart_data(emotions)
     chart_data_json = JsonResponse(data, safe=False).content.decode('utf-8')
